chore: rearrange and cleanup
