+++ crates/node_binding/src/js_values/mod.rs
@@ -1,0 +1,7 @@
+mod asset;
+mod plugin;
+mod stats;
+
+pub use asset::*;
+pub use plugin::*;
+pub use stats::*;
